lib: faster event listening
The current EventEmitter uses arrays to control multiple listeners,
but that means that prependListener and removeListeners operations
have an O(N) time complexity. With this change, instead of an array
we use a linkedList + a map, which gives an O(1) time complexity
for those operations. This will be speciall helpful in scenarios
where we have many listeners to the same event.

BREAKING CHANGE: Now, the properties of _events will no longer be
a function or an array, but a function or a "MappedLinkedList". Some
of the operations from the array are present, though, like
[Symbol.iterator], unshift and push, making possible to adapt some
code that are arbritary using it